fix: waiver revoke — ambiguous operator on unresolved parameter types (#49)
Root cause of the "Failed to revoke waiver" 500, confirmed via the
temporary error-surfacing change: Postgres error "operator is not
unique: unknown + unknown".

WHEN $1 + $2 >= amount_due — neither $1 nor $2 has a directly-typed
operand within that specific sub-expression (SET waiver_amount = $1 and
amount_paid = $2 assign them to typed columns elsewhere in the same
statement, but that didn't carry over to disambiguate this arithmetic
expression), so Postgres couldn't pick between its overloaded '+'
operators for two parameters of type "unknown".

Cast both explicitly: $1::numeric + $2::numeric. Checked the rest of the
fee routes for the same bare-parameter-arithmetic pattern — the only
other lookalike (PATCH's COALESCE(waiver_amount, 0) + $1) already has one
typed operand from the column reference, so it wasn't affected.

+++ app/api/fees/waivers/route.ts
@@ -305,8 +305,8 @@ async function handleDELETE(req: NextRequest) {
          SET waiver_amount = $1,
              amount_paid   = $2,
              status = CASE
-               WHEN $1 + $2 >= amount_due THEN (CASE WHEN $1 > 0 THEN 'waived' ELSE 'paid' END)
-               WHEN $2 > 0 THEN 'partial'
+               WHEN $1::numeric + $2::numeric >= amount_due THEN (CASE WHEN $1::numeric > 0 THEN 'waived' ELSE 'paid' END)
+               WHEN $2::numeric > 0 THEN 'partial'
                WHEN EXISTS (SELECT 1 FROM academic_years ay WHERE ay.school_id = school_id AND ay.label = academic_year AND ay.end_date < CURRENT_DATE) THEN 'overdue'
                ELSE 'pending'
              END
